requirements.txt
pandas
numpy
seaborn
pyyaml
-e .